Add appointments to dataset definition
Following the logic of f2f and virtual consultations, this creates one variable for the count of appointments and one logical variable that indicates whether or not someone has had an appointment. This currently follows includes some types of appointments and we should review if this is appropriate for this study.