Count appts in measures only when registered  and age >= 18
This is an important addition for the measures that use count_appointment as the denominator. These measures now also only include patients with a registration and an age of 18 or above. The other measures that use count_appointment as the numerator will stay the same as both of these conditions also appear in their denominators.